requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# astronomer-cosmos[dbt-bigquery]==1.0.3
# protobuf==3.20.0
# apache-airflow-providers-dbt-cloud==1.0.2
# polygon-api-client==1.10.1
apache-airflow-providers-google==10.3.0
faker==19.6.2
google-cloud-pubsub==2.17.0
# soda-core-bigquery==3.0.45